5/11 board/details부분 겹침
