+++ TProject/src/main/java/com/project/controller/HomeController.java
@@ -349,119 +349,5 @@ public class HomeController {
 		return "/upload/" + newFileName;
 	}
 	
-	@RequestMapping(value="/jlist.do",method=RequestMethod.GET)
-	public String jList(@RequestParam Map<String, Object> params, Model model,HttpServletRequest request) {
-		if(params.get("bIdx") == null) {
-			params.put("bIdx", 3);
-		}
-		System.out.println(params.toString());
-		List<ArticlesVO> list = boardService.Jlist(params,request);
-		
-		if( list.size() >0) {
-			
-		Map<Integer,String> map=new HashMap<Integer, String>();
-			for(int i=0;i<list.size();i++) {
-				String co=list.get(i).getContent().replaceAll(" ", "");
-				
-				
-				if(co.indexOf("<img") > -1) {					
-				int startIndex=co.indexOf("<img");
-				startIndex=co.indexOf("src", startIndex);
-				String co2=co.substring(startIndex+5);
-				String[] co3=co2.split("\"");
-				String url=co3[0];				
-				map.put(i, url);
-				} else {
-					map.put(i, "");
-				}
-				model.addAttribute("imgsrc", map);
-				
-				System.out.println(map.toString());
-			}
-		}
-		
-		int articlesTotal=(Integer)request.getAttribute("count");
-		System.out.println("총게시물"+articlesTotal);
-		
-		//카멜?기법
-		model.addAttribute("list", list);
-		model.addAttribute("status", params.get("status"));
-		model.addAttribute("searchtitle", params.get("searchtitle"));
-		model.addAttribute("articlesTotal",articlesTotal);
-		
-		return "/board/jlist";
-		
-		
-	}
 	
-		@RequestMapping(value="jlist/detail.do", method=RequestMethod.GET)
-		public String jlistView(Model model,ArticlesVO vo,@RequestParam Map<String, Object> params,HttpServletRequest request) {
-		System.out.println(params.get("bIdx"));
-		System.out.println(params.get("aIdx"));
-			
-		boardService.readCount(vo);
-		
-		Map<String, Object> one = boardService.jlistOneArticle(params,request);
-		
-		GeneralMembersVO writer=new GeneralMembersVO();
-		int a=(int) one.get("mIdx");
-		writer.setmIdx(a);
-		
-		writer=memberService.oneMemberInfo(writer);
-		
-		
-		System.out.println("게시글정보"+one.toString());
-		
-		model.addAttribute("cmtCount",request.getAttribute("cmtCount"));//댓글총개수?
-		model.addAttribute("vo",one);//게시글정보 보내기
-		model.addAttribute("profileSrc",writer.getProfileSrc());//글 작성자 프로필 사진 
-		
-		return "/board/jlist/details";
-	}
-		
-		@RequestMapping(value="jlist/commentWrite.do")
-		@ResponseBody
-		public int commentWrite(CommentsVO vo,@RequestParam("commentSrc") MultipartFile file,HttpServletRequest request) throws IllegalStateException, IOException {
-					
-			
-			if(vo.getContent() != null && vo.getContent() != "") {
-				
-				if(!file.isEmpty()) {
-					System.out.println(file);
-					
-					String path = request.getSession().getServletContext().getRealPath("/resources/upload"); //실제경로		
-						
-					String fileName=file.getOriginalFilename();
-					
-					String extension = fileName.substring(fileName.lastIndexOf("."), fileName.length());
-					
-					UUID uuid = UUID.randomUUID();
-					
-					String newFileName = uuid.toString() + extension;
-		
-					File target = new File(path, newFileName);
-					
-					System.out.println(target.toString());
-					
-					file.transferTo(target);//파일이생성됨
-					
-					vo.setPicSrc("/upload/"+newFileName);
-				}
-			
-			return boardService.commentWrite(vo);
-			
-			} else {
-				
-				return -1;
-			}
-		
-		}
-		
-		@RequestMapping(value="jlist/commentList.do")
-		@ResponseBody
-		public List<CommentsVO> commentsList(@RequestParam Map<String, Object> params){
-			
-			
-			return boardService.commentList(params);
-		}
 }
