+++ TProject/src/main/java/com/project/service/BoardService.java
@@ -31,6 +31,4 @@ public interface BoardService {
 	Map<String,Object> jlistOneArticle(Map<String, Object> map,HttpServletRequest request);
 	int commentWrite(CommentsVO vo);
 	List<CommentsVO> commentList(Map<String, Object> params);
-	int insertComments(CommentsVO vo);
-	List<CommentsVO> cList(CommentsVO vo);
 }
